Restructure code and configurations.
- histograms and plotters are moved to the relevant python module (in progress)
- configuration parameters are slimmed and with more meaningful names
- printouts are added to ease understanding of plotter configuration
