+++ cfg/eg_rate.yaml
@@ -1,0 +1,35 @@
+
+common:
+  output_dir:
+    default: /eos/user/c/cerminar/hgcal/CMSSW1015/plots/
+    triolet: /Users/cerminar/cernbox/hgcal/CMSSW1015/plots/
+  plot_version: v160A
+  run_clustering: False
+  run_density_computation: False
+  file_label: egrate
+
+
+rate_ctl2:
+  plotters:
+    # - !!python/name:cfg.eg_rate.eg_emuCTl1_sta_plotters
+    # - !!python/name:cfg.eg_rate.eg_emuCTl1_ele_plotters
+    # - !!python/name:cfg.eg_rate.eg_emuCTl1_pho_plotters
+    - !!python/name:cfg.eg_rate.eg_emuCTl2_plotters
+    # - !!python/name:cfg.eg_rate.eg_emuCTl1_ell_plotters
+    # - !!python/name:cfg.eg_rate.eg_emuCTl2_ell_plotters
+
+rate_menu:
+  file_label:
+    menu
+  plotters:
+    # - !!python/name:cfg.eg_rate.eg_emu_plotters
+    - !!python/name:cfg.eg_rate.eg_menuSta_plotters
+    - !!python/name:cfg.eg_rate.eg_menuCTl2_plotters
+    # - !!python/name:cfg.eg_rate.eg_menuCTl2_ell_plotters
+
+rate_counter_menu:
+  file_label: count
+  plotters:
+    # - !!python/name:cfg.eg_rate.eg_emu_plotters
+    - !!python/name:cfg.eg_rate.eg_menuCTl2_rate
+
